% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sampling_frequency}
\alias{sampling_frequency}
\title{Sampling fequency}
\format{
A data frame:
\describe{
\item{population_group}{Population group}
\item{number_sampled}{Number of individuals sampled per population group}
\item{number_population}{Estimated number of individuals in each population group}
}
}
\source{
\url{https://magosil86.github.io/bumblebee/}
}
\usage{
sampling_frequency
}
\description{
Estimated number of individuals with HIV in intervention and control communities
of the BCPP/Ya Tsie trial, and the number of individuals sampled from each for
HIV viral phylogenetic analysis.
}
\references{
Magosi LE, et al., Deep-sequence phylogenetics to quantify patterns of
HIV transmission in the context of a universal testing and treatment
trial – BCPP/ Ya Tsie trial. To submit for publication, 2021.
}
\keyword{datasets}
