% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{counts_hiv_transmission_pairs}
\alias{counts_hiv_transmission_pairs}
\title{Observed HIV transmission pairs}
\format{
A data frame:
\describe{
\item{H1_group}{Name of population group 1}
\item{H2_group}{Name of population group 1}
\item{num_linked_pairs_observed}{Number of observed directed transmission pairs between samples from population groups 1 and 2}}
}
\source{
\url{https://magosil86.github.io/bumblebee}
}
\usage{
counts_hiv_transmission_pairs
}
\description{
Counts of directed HIV transmission pairs observed within and between
intervention and control communities in the 30-community BCPP/Ya Tsie
HIV prevention trial in Botswana (2013-2018). The Botswana -Ya Tsie
trial was a pair-matched community randomized trial that evaluated
the effect of a universal HIV test and treat intervention in reducing
population-level incidence. For further details see:
\url{https://magosil86.github.io/bumblebee}.
}
\references{
Makhema, J., et al., Universal Testing, Expanded Treatment, and Incidence
of HIV Infection in Botswana. New England Journal of Medicine, 2019.
381(3): p. 230-242.
}
\keyword{datasets}
